requirements.txt
Pillow
pytest
codecov
pytest-cov
pytest-pylint
typeguard
